+++ src/App.css
@@ -43,7 +43,7 @@
 }
 
 .error-message {
-  color: red;
+  color: #d32f2f;
 }
 
 .success-message {
