+++ src/components/feedBack/feedback.css
@@ -32,4 +32,3 @@
   color: #757575;
   font-style: italic;
 }
-
